// src/app/models/lastfm.ts
export type LastFMParam = string | string[];
export type LastFMParams<T> = Record<string, LastFMParam | T>;
export type LastFMVoidOrNumber = void | number;
export type LastFMBooleanNumber = 0 | 1;
export type LastFMUnknownFunction = (...args: unknown[]) => unknown;
export type LastFMRequestParams<T> = Record<string, LastFMParam | T>;
// ...
export type LastFMUserParams = LastFMRequestParams<
  LastFMParam | number | void | boolean
> &
  Readonly<{
    user: string;
  }>;

export type LastFMUserOptionalParams = Readonly<{
  page?: number;
  limit?: number;
}>;
// ...
export type LastFMUserGetRecentTracksParams = LastFMUserParams &
  LastFMUserOptionalParams &
  Readonly<{
    to?: string;
    from?: string;
    extended?: 0 | 1;
  }>;
// ...
export type LastFMUserGetInfoResponse = Readonly<{
  user: {
    name: string;
    age: string;
    subscriber: LastFMBooleanNumber;
    realname: string;
    bootstrap: LastFMBooleanNumber;
    playcount: string;
    artist_count: string;
    playlists: string;
    track_count: string;
    album_count: string;
    image: Array<{
      '#text': string;
      size: string;
    }>;
    registered: {
      unixtime: string;
      '#text': number;
    };
    country: string;
    gender: string;
    url: string;
    type: string;
  };
}>;
// ...
export type LastFMUserGetRecentTracksResponse = Readonly<{
  recenttracks: {
    track: Array<{
      artist: {
        mbid: string;
        '#text': string;
      };
      streamable: LastFMBooleanNumber;
      image: Array<{
        '#text': string;
        size: string;
      }>;
      mbid: string;
      album: {
        mbid: string;
        '#text': string;
      };
      name: string;
      url: string;
      date: {
        uts: string;
        '#text': string;
      };
      '@attr'?: {
        nowplaying: 'true';
      };
    }>;
    '@attr': {
      user: string;
      totalPages: string;
      page: string;
      perPage: string;
      total: string;
    };
  };
}>;